Move RoadRemovalArgs to RemoveRoad.h
