+++ src/OpenLoco/src/Windows/Construction/ConstructionTab.cpp
@@ -3,6 +3,7 @@
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/GameCommands.h"
 #include "GameCommands/Road/CreateRoad.h"
+#include "GameCommands/Road/RemoveRoad.h"
 #include "GameState.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
